software/requirements.txt
numpy
picamera2
pillow
pynput
